+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - New contour integral method in `DSNEP`.
 - New function `BVSetRandomSign()` for random values equal to -1 or 1 with equal probability.
 - New functions `RGIsAxisymmetric()`, `RGCanUseConjugates()` and `RGComputeQuadrature()`.
+- New function `SlepcHasExternalPackage()` to check whether SLEPc was configured with the
+  given external package.
 
 ### Changed